backend/requirements.txt
fastapi
uvicorn 
diffusers
transformers
accelerate
safetensors
omegaconf
pytorch_lightning


